Rename components for consistency and restructure layout files
Refactored the naming of various components for improved clarity and coherence. Moved `Layout` component and associated styles into a dedicated directory to better encapsulate layout-related files. Adjusted import statements across the project to reflect the new structure. This reorganization simplifies navigation and maintenance of the component hierarchy.
