+++ src/components/index.ts
@@ -1,11 +1,11 @@
 export * from './ExternalLink';
-export * from './oldheader';
-export * from './oldimageHeader';
-export * from './oldlayout';
+export * from './Header';
+export * from './ImageHeader';
+export * from './Layout';
 export * from './ListTags';
 export * from './PassWords';
 export * from './CodePen';
-export * from './oldposts';
-export * from './oldseo';
-export * from './oldsitemap';
+export * from './Posts';
+export * from './Seo';
+export * from './Sitemap';
 export * from './Tags';
