+++ src/pages/404.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 
-import { Layout, Seo, OldSiteMap } from '../components/';
+import { Layout, Seo, Sitemap } from '../components/';
 
 const NotFoundPage = () => (
   <Layout>
     <h1>404: Not Found</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
     <h2>Maybe you're looking for one of these:</h2>
-    <OldSiteMap />
+    <Sitemap />
   </Layout>
 );
